Rename module back to auth
