+++ auth/main.py
@@ -11,7 +11,7 @@ app.config.from_object(app_settings)
 config = app.config
 db = SQLAlchemy(app)
 
-from token_service.api import auth_blueprint
+from auth.api import auth_blueprint
 
 app.register_blueprint(auth_blueprint, url_prefix="/auth")
